move stringHasValue into utilities/guards
