+++ src/utilities/guards.spec.ts
@@ -1,5 +1,5 @@
-import { describe, expect, test } from 'vitest'
-import { hasProperty } from '@/utilities/guards'
+import { describe, expect, expectTypeOf, test } from 'vitest'
+import { hasProperty, StringHasValue } from '@/utilities/guards'
 
 describe('hasProperty', () => {
 
@@ -27,4 +27,41 @@ describe('hasProperty', () => {
     expect(source).toBe(false)
   })
 
+})
+
+describe('stringHasValue', () => {
+  test('given empty string, returns false', () => {
+  type Source = StringHasValue<''>
+  type Expect = false
+
+  expectTypeOf<Source>().toEqualTypeOf<Expect>()
+  })
+
+  test('given generic string, returns true', () => {
+  type Source = StringHasValue<string>
+  type Expect = true
+
+  expectTypeOf<Source>().toEqualTypeOf<Expect>()
+  })
+
+  test('given type any, returns true', () => {
+  type Source = StringHasValue<any>
+  type Expect = true
+
+  expectTypeOf<Source>().toEqualTypeOf<Expect>()
+  })
+
+  test('given type other than string or any, returns false', () => {
+  type Source = StringHasValue<undefined>
+  type Expect = false
+
+  expectTypeOf<Source>().toEqualTypeOf<Expect>()
+  })
+
+  test('given string not empty, returns true', () => {
+  type Source = StringHasValue<'foo'>
+  type Expect = true
+
+  expectTypeOf<Source>().toEqualTypeOf<Expect>()
+  })
 })
