+++ src/utilities/index.ts
@@ -3,5 +3,4 @@ export * from './checkDuplicateKeys'
 export * from './guards'
 export * from './isBrowser'
 export * from './makeOptional'
-export * from './string'
 export * from './testHelpers'
